streaming-lakehouse/producer/requirements.txt
kafka-python
Faker
